Initial commit
Added the main.py file. The main.py file is the main script for the application. It contains the entire code base for the proper functioning of the application. The documentation of the application might not be of that excellence, you can pardon me on that behalf that this application is prepared to be submitted as a high school computer programming project.